Move New submission form to submission folder
